pyproject.toml
[tool.poetry]
name = "carry"
version = "0.1.0"
description = "The telegram bot for bonus program"
authors = ["o.kyba <vestigor0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.1"
dynaconf = "^3.1.12"
jinja2 = "^3.1.2"
qrcode = "^7.4.2"
sqlalchemy = "^2.0.5.post1"
asyncpg = "^0.27.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
watchdog = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py311']
skip-string-normalization = 1

[tool.isort]
py_version = "39"
profile = "black"
multi_line_output = 3
line_length = 79
length_sort = true
include_trailing_comma = true
